wasm-smith: Move core Wasm generation into a `core` submodule
