+++ crates/wasm-smith/src/core/code_builder.rs
