+++ crates/wasm-smith/src/core/encode.rs
