+++ crates/wasm-smith/src/core/terminate.rs
